Fixed > 8 textures on terrains
So turns out you cannot store integers in a shader graph so > 16 bits would break for the compression. So i would need to seperate vector3s into multiple floats * 2 for splitting into 16bit halfs. so easier solution was to write those to a texture and pass that into the shader. my brain is dead

!! Dont forget to change the shader layer index to be how many are actually assigned

- Fixed texture arrays being scrambled after assigning a 9th texture
- Fixed textures white when having > 8 textures
- Now passing in array assigned textures texture to the shader